backend/app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
pdfminer.six==20250327
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
python-docx==1.1.2
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
